Move encryption to separate file
