+++ full-service/src/db/mod.rs
@@ -5,49 +5,33 @@
 pub mod account;
 pub mod account_txo_status;
 pub mod assigned_subaddress;
+pub mod encryption;
 pub mod encryption_indicator;
 pub mod models;
 pub mod schema;
 pub mod transaction_log;
 pub mod txo;
 
-use self::{
-    encryption_indicator::{EncryptionModel, EncryptionState},
-    models::EncryptionIndicator,
-};
+use self::encryption::EncryptionProvider;
 use crate::error::WalletDbError;
 
 use mc_account_keys::PublicAddress;
-use mc_common::logger::{log, Logger};
+use mc_common::logger::Logger;
 
-use aes_gcm::{
-    aead::{
-        generic_array::{sequence::Split, GenericArray},
-        Aead, AeadInPlace, NewAead,
-    },
-    Aes256Gcm,
-};
-use blake2::{Blake2b, Digest};
 use diesel::{
     prelude::*,
     r2d2::{ConnectionManager, Pool, PooledConnection},
 };
-use std::{
-    convert::TryFrom,
-    sync::{Arc, Mutex},
-};
+use std::{convert::TryFrom, sync::Arc};
 
-/// Domain tag for database-wide encryption.
-pub const ENCRYPTION_KEY_DOMAIN_TAG: &str = "mc_full_service";
-const ENCRYPTION_VERIFICATION_VAL: &[u8] = b"verify";
-
-// Helper method to use our PrintableWrapper to b58 encode the PublicAddress
+/// Helper method to use our PrintableWrapper to b58 encode the PublicAddress
 pub fn b58_encode(public_address: &PublicAddress) -> Result<String, WalletDbError> {
     let mut wrapper = mc_mobilecoind_api::printable::PrintableWrapper::new();
     wrapper.set_public_address(public_address.into());
     Ok(wrapper.b58_encode()?)
 }
 
+/// Helper method to use our PrintableWrapper when decoding a b58 PublicAddress
 pub fn b58_decode(b58_public_address: &str) -> Result<PublicAddress, WalletDbError> {
     let wrapper =
         mc_mobilecoind_api::printable::PrintableWrapper::b58_decode(b58_public_address.to_string())
@@ -63,15 +47,18 @@ pub fn b58_decode(b58_public_address: &str) -> Result<PublicAddress, WalletDbErr
     Ok(PublicAddress::try_from(pubaddr_proto).unwrap())
 }
 
-type ExpandedPassword = (
-    GenericArray<u8, <Aes256Gcm as NewAead>::KeySize>,
-    GenericArray<u8, <Aes256Gcm as AeadInPlace>::NonceSize>,
-);
+/// A struct encapsulating a single connection, which can also be used for encryption and logging.
+/// Intended to be passed to trait methods in various db::models.
+pub struct WalletDbConn {
+    pub conn: PooledConnection<ConnectionManager<SqliteConnection>>,
+    pub encryption_provider: Arc<EncryptionProvider>,
+    pub logger: Logger,
+}
 
 #[derive(Clone)]
 pub struct WalletDb {
     pool: Pool<ConnectionManager<SqliteConnection>>,
-    password_hash: Arc<Mutex<Vec<u8>>>,
+    encryption_provider: EncryptionProvider,
     logger: Logger,
 }
 
@@ -79,7 +66,7 @@ impl WalletDb {
     pub fn new(pool: Pool<ConnectionManager<SqliteConnection>>, logger: Logger) -> Self {
         Self {
             pool,
-            password_hash: Arc::new(Mutex::new(vec![])),
+            encryption_provider: EncryptionProvider::new(logger.clone()),
             logger,
         }
     }
@@ -99,68 +86,30 @@ impl WalletDb {
         Ok(self.pool.get()?)
     }
 
-    /// Set the password hash for the DB.
+    /// Set the password hash for the database.
     pub fn set_password_hash(
         &self,
         password_hash: &[u8],
         conn: &PooledConnection<ConnectionManager<SqliteConnection>>,
     ) -> Result<(), WalletDbError> {
-        {
-            let mut cur_password_hash = self.password_hash.lock()?;
-            *cur_password_hash = password_hash.to_vec();
-        }
-        // Encrypt the verification value and set in the DB
-        let verification_value = Self::encrypt(ENCRYPTION_VERIFICATION_VAL, password_hash)?;
-        EncryptionIndicator::set_verification_value(&verification_value, conn)?;
-
-        Ok(())
+        Ok(self
+            .encryption_provider
+            .set_password_hash(password_hash, conn)?)
     }
 
+    /// Get the password hash for the database.
     pub fn get_password_hash(&self) -> Result<Vec<u8>, WalletDbError> {
-        let cur_password_hash = self.password_hash.lock()?;
-        Ok(cur_password_hash.clone())
+        Ok(self.encryption_provider.get_password_hash()?)
     }
 
+    /// Check whether the database is currently unlocked.
     pub fn is_unlocked(&self) -> Result<bool, WalletDbError> {
-        Ok(!self.password_hash.lock()?.is_empty())
+        Ok(self.encryption_provider.is_unlocked()?)
     }
 
     pub fn unlock(&self, password_hash: &[u8]) -> Result<(), WalletDbError> {
-        // No need to check db state if we're already unlocked
-        if self.is_unlocked()? {
-            return Ok(());
-        }
-
         let conn = self.get_conn()?;
-        // Check whether encrypted, and if so, then attempt to unlock
-        match EncryptionIndicator::get_encryption_state(&conn)? {
-            EncryptionState::Empty => {
-                log::info!(
-                    self.logger,
-                    "DB has never been locked. Please call set_password to enable encryption."
-                );
-                return Err(WalletDbError::SetPassword);
-            }
-            EncryptionState::Encrypted => {
-                log::debug!(self.logger, "DB is locked. Verifying password.");
-                // Attempt to decrypt the test value to confirm if password is correct
-                let expected_val = Self::encrypt(ENCRYPTION_VERIFICATION_VAL, password_hash)?;
-                if EncryptionIndicator::verify_password(&expected_val, &conn)? {
-                    // Store password hash in memory
-                    let mut cur_password_hash = self.password_hash.lock()?;
-                    *cur_password_hash = password_hash.to_vec();
-                } else {
-                    return Err(WalletDbError::PasswordFailed);
-                }
-            }
-            EncryptionState::Unencrypted => {
-                log::info!(
-                    self.logger,
-                    "DB is unencrypted. Please call set_password to enable encryption."
-                );
-            }
-        }
-        Ok(())
+        Ok(self.encryption_provider.unlock(password_hash, &conn)?)
     }
 
     pub fn change_password(
@@ -169,88 +118,25 @@ impl WalletDb {
         new_password_hash: &[u8],
     ) -> Result<(), WalletDbError> {
         let conn = self.get_conn()?;
-        // Check whether encrypted, and if so, then verify old password. For any other state, set_password to new.
-        match EncryptionIndicator::get_encryption_state(&conn)? {
-            EncryptionState::Empty => {
-                log::info!(
-                    self.logger,
-                    "Database has never been locked. Setting password with new password."
-                );
-                self.set_password_hash(new_password_hash, &conn)?;
-            }
-            EncryptionState::Encrypted => {
-                log::debug!(self.logger, "Database is locked. Verifying old password.");
-                // Attempt to decrypt the test value to confirm if password is correct
-                let expected_val = Self::encrypt(ENCRYPTION_VERIFICATION_VAL, old_password_hash)?;
-                if EncryptionIndicator::verify_password(&expected_val, &conn)? {
-                    // Set password to new password
-                    self.set_password_hash(new_password_hash, &conn)?;
-                } else {
-                    return Err(WalletDbError::PasswordFailed);
-                }
-            }
-            EncryptionState::Unencrypted => {
-                log::info!(
-                    self.logger,
-                    "Database is unencrypted. Setting password with new password."
-                );
-                self.set_password_hash(new_password_hash, &conn)?;
-            }
-        }
-        Ok(())
-    }
-
-    /// Encrypt data.
-    pub fn encrypt(plaintext_bytes: &[u8], password_hash: &[u8]) -> Result<Vec<u8>, WalletDbError> {
-        let (key, nonce) = Self::expand_password(&password_hash)?;
-        let cipher = Aes256Gcm::new(&key);
-        Ok(cipher.encrypt(&nonce, &plaintext_bytes[..])?)
-    }
-
-    /// Decrypt data.
-    pub fn decrypt(ciphertext: &[u8], password_hash: &[u8]) -> Result<Vec<u8>, WalletDbError> {
-        let (key, nonce) = Self::expand_password(&password_hash)?;
-        let cipher = Aes256Gcm::new(&key);
-        Ok(cipher.decrypt(&nonce, ciphertext)?)
-    }
-
-    /// Expands the password into an encryption key and a nonce.
-    fn expand_password(password: &[u8]) -> Result<ExpandedPassword, WalletDbError> {
-        // Hash the password hash with Blake2b to get 64 bytes, first 32 for aeskey, second 32 for nonce
-        let mut hasher = Blake2b::new();
-        hasher.update(&ENCRYPTION_KEY_DOMAIN_TAG);
-        hasher.update(&password);
-        let result = hasher.finalize();
-
-        let (key, remainder) = Split::<u8, <Aes256Gcm as NewAead>::KeySize>::split(result);
-        let (nonce, _remainder) =
-            Split::<u8, <Aes256Gcm as AeadInPlace>::NonceSize>::split(remainder);
-
-        Ok((key, nonce))
+        Ok(self
+            .encryption_provider
+            .change_password(old_password_hash, new_password_hash, &conn)?)
     }
 }
 
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::test_utils::WalletDbTestContext;
+    use crate::{
+        db::{
+            encryption_indicator::{EncryptionModel, EncryptionState},
+            models::EncryptionIndicator,
+        },
+        test_utils::WalletDbTestContext,
+    };
     use mc_common::logger::{test_with_logger, Logger};
     use mc_crypto_rand::rand_core::RngCore;
     use rand::{rngs::StdRng, SeedableRng};
-
-    // Encrypting and decrypting with a set password should succeed.
-    #[test]
-    fn test_basic_encrypt_decrypt() {
-        let mut rng: StdRng = SeedableRng::from_seed([20u8; 32]);
-
-        let mut password_hash = [1u8; 32];
-        rng.fill_bytes(&mut password_hash);
-
-        let plaintext = b"test_plaintext";
-        let ciphertext = WalletDb::encrypt(plaintext, &password_hash).unwrap();
-        let decrypted = WalletDb::decrypt(&ciphertext, &password_hash).unwrap();
-        assert_eq!(plaintext.to_vec(), decrypted);
-    }
 
     // Initializing DB, setting password, and unlocking should behave as expected.
     #[test_with_logger]
